requirements.txt
contourpy==1.2.0
cycler==0.12.1
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.3
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.8.2
six==1.16.0
